AppVeyorPoc: Component oriented disign with new customization rules.
