--- Source/Presentation/Web/AppVeyorPoc/src/app-layer1/projects/projects.component.css
@@ -1,5 +1,0 @@
-﻿@import '../../app/projects/projects.component.css';
-
-li{
-  background-color: rebeccapurple;
-}
