--- Source/Presentation/Web/AppVeyorPoc/src/app-layer1/projects/projects.component.html
@@ -1,7 +1,0 @@
-﻿<h1>List of projects</h1>
-<h2>Layer1 influence</h2>
-<ul>
-  <li *ngFor="let project of projects">
-    <a [routerLink]="['/project', project.name]">{{project.name}}</a>
-  </li>
-</ul>
