--- Source/Presentation/Web/AppVeyorPoc/src/app-layer2/projects/projects.component.css
@@ -1,5 +1,0 @@
-﻿@import '../../app-layer1/projects/projects.component.css';
-
-li{
-  background-color: burlywood;
-}
