--- Source/Presentation/Web/AppVeyorPoc/src/app-layer3/styles.css
@@ -1,3 +1,0 @@
-﻿h1 {
-  color: green;
-}
