+++ Source/Presentation/Web/AppVeyorPoc/src/app/app.component.ts
@@ -2,7 +2,8 @@
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html'
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.css']
 })
 
 export class AppComponent {
