+++ Source/Presentation/Web/AppVeyorPoc/src/app/global.css
@@ -1,0 +1,7 @@
+﻿h1 {
+  color: brown;
+}
+
+h2 {
+  color: blue;
+}
