+++ Source/Presentation/Web/AppVeyorPoc/src/app/global.custom.css
@@ -1,2 +1,2 @@
 ﻿/* Use imported file for you global style customisation */
-@import 'app/global.custom.css';
+@import 'global.css';
