+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/artifacts/artifacts.component.css
@@ -1,0 +1,3 @@
+﻿h4 {
+  color: magenta;
+}
