+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/artifacts/artifacts.component.html
