+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/artifacts/artifacts.component.ts
@@ -1,0 +1,13 @@
+﻿import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: './artifacts.component.html',
+  styleUrls: ['./artifacts.component.css']
+})
+
+export class ArtifactsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void { }
+}
