+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/console/console.component.html
