+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/console/console.component.ts
@@ -1,0 +1,13 @@
+﻿import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: './console.component.html',
+  styleUrls: ['./console.component.css']
+})
+
+export class ConsoleComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void { }
+}
