+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/messages/messages.component.css
@@ -1,0 +1,3 @@
+﻿h4{
+  color: rebeccapurple;
+}
