+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build-area/messages/messages.component.ts
@@ -1,0 +1,13 @@
+﻿import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: './messages.component.html',
+  styleUrls: ['./messages.component.css']
+})
+
+export class MessagesComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void { }
+}
