+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/build.ts
