+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/history/history.component.html
