+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/project-details/project-details.component.css
@@ -1,0 +1,1 @@
+﻿
