+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/project-details/project-details.component.html
