+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/project-details/project-details.custom-component.css
@@ -1,0 +1,5 @@
+﻿@import 'project-details.component.css';
+
+label{
+  background-color: sienna;
+}
