+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/project.ts
