+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/projects/confirmation-popup.html
@@ -1,0 +1,3 @@
+﻿<label>Are you sure?</label>
+<button>Yes</button>
+<button>No</button>
