+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/projects/projects.component.css
@@ -1,0 +1,3 @@
+﻿ul{
+  background-color: forestgreen;
+}
