+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/projects/projects.component.html
