+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/projects/projects.custom-component.css
@@ -1,0 +1,5 @@
+﻿@import 'projects.component.css';
+
+li{
+  background-color: darkorange;
+}
