+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/projects/projects.custom-component.html
@@ -1,0 +1,6 @@
+﻿<h1>List of projects(custom)</h1>
+<ul>
+  <li *ngFor="let project of projects">
+    <a [routerLink]="['/project', project.name]">{{project.name}}</a>
+  </li>
+</ul>
