+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/settings/settings.component.html
