+++ Source/Presentation/Web/AppVeyorPoc/src/app/projects-area/settings/settings.component.ts
@@ -1,0 +1,13 @@
+﻿import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: './settings.component.html',
+  styleUrls: ['./settings.component.css']
+})
+
+export class SettingsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void { }
+}
