--- Source/Presentation/Web/AppVeyorPoc/src/app/projects/deployments.component.ts
@@ -1,13 +1,0 @@
-﻿import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-deployments',
-  templateUrl: './deployments.component.html'
-})
-
-export class DeploymentsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void { }
-}
