--- Source/Presentation/Web/AppVeyorPoc/src/app/projects/projects.component.css
@@ -1,3 +1,0 @@
-﻿ul{
-  background-color: aquamarine;
-}
