--- Source/Presentation/Web/AppVeyorPoc/src/app/projects/settings.component.ts
@@ -1,13 +1,0 @@
-﻿import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-settings',
-  templateUrl: './settings.component.html'
-})
-
-export class SettingsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void { }
-}
