--- Source/Presentation/Web/AppVeyorPoc/src/app/styles.css
@@ -1,7 +1,0 @@
-﻿h1 {
-  color: red;
-}
-
-h2{
-  color: blue;
-}
